[COLUMNAR UI] MB-65179 - Unabled to edit RBAC group from UI
In the Cluster -> Security pane, the edit button for Groups was not
working. The underlying problem was that, in morpheus, the Add/Edit
Group dialog had been changed to require `permissions` to be
specified. This was updated in the call to "Add Group", but not in the
call where an existing Group is edited.

I also removed the unused addRolesGroup function, since that
functionality is actually taken care of in mn_roles_controller.

Change-Id: Ic0b27020c1075a6bc3cc1ac2dda3642e6ca5fde3
Reviewed-on: https://review.couchbase.org/c/ns_server/+/224206
Reviewed-on: https://review.couchbase.org/c/columnar-ui-v1/+/224553
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Michael Blow <michael.blow@couchbase.com>