[COLUMNAR UI] MB-64871 Encryption At Rest Fixes
Times displayed in Encryption at Rest pages are in local time
with the GMT clearly specified when setting the rotation time
of an encryption key.
Added limits for hour and minutes fields.
24 hour format clarification when setting the rotation time
of an encryption key.
Smaller fields for hour and minute when setting
the rotation time of an encryption key.

Change-Id: I7fafa94389f3dbbd47f248c98bc8d32dada88de8
Original-Review: https://review.couchbase.org/c/ns_server/+/224489
Original-Author: lukaisthewolf <raluca.lupu@couchbase.com>
Original-Commit: 331e1acb8
Reviewed-on: https://review.couchbase.org/c/columnar-ui-v1/+/224920
Tested-by: Michael Blow <michael.blow@couchbase.com>
Reviewed-by: Michael Blow <michael.blow@couchbase.com>