[COLUMNAR UI] MB-50503 Show Local Time Across UI
The UI shows the date times in local time across the app.
A UTC offset text and a tooltip are added
for extra explanations in the UI
   - local UTC offset
   - server UTC offset
This change integrates the API changes related to server times
which now include the timezone shifts.
UI transforms the API server times in the user local time
(using the timezone shift received from the API and the user's
local timezone).
The API change is available for cluster compatibility
>= 8.0.

Change-Id: Ic7a763d9a86c9992e95c6dff342616c740db6154
Original-Review: https://review.couchbase.org/c/ns_server/+/224625
Original-Author: lukaisthewolf <raluca.lupu@couchbase.com>
Original-Commit: f36e0b3b8
Reviewed-on: https://review.couchbase.org/c/columnar-ui-v1/+/226796
Tested-by: Michael Blow <michael.blow@couchbase.com>
Reviewed-by: Michael Blow <michael.blow@couchbase.com>